% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_graph.R
\name{aa_graph}
\alias{aa_graph}
\title{aa_graph}
\usage{
aa_graph(aa_seq)
}
\arguments{
\item{aa_seq}{amino acid sequence}
}
\value{
aa_plot plot of the amino acid frequencies
}
\description{
aa_graph
}
\examples{
If you input LLNRKKK, L is 2/7; N is 1/7, R is 1/7 and K is 3/7
}
