% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{codon_table}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
This is a table with the DNA codes for the 20 amino acids. There are 20 letters, 1 per amino acid and 64 possible nucleotide combinations.
}
\keyword{datasets}
