% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_create.R
\name{dna_create}
\alias{dna_create}
\title{Create DNA strand}
\usage{
dna_create(dna_length)
}
\arguments{
\item{dna_length}{int that indicates the length of the desired DNA strand.}
}
\value{
dna_seq string of nucleotides with the desired length
}
\description{
Create DNA strand
}
\examples{
dna_length = 8, dna_seq = "ACTCAGGT"
}
