% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_sequence.R
\name{get_random_sequence}
\alias{get_random_sequence}
\title{Get a random sequnce}
\usage{
get_random_sequence(amount)
}
\arguments{
\item{amount}{amount of bases in your random sequnce}
}
\value{
a random sequence of a specified amount of bases
}
\description{
Get a random sequnce
}
