% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_codons.R
\name{order_codons}
\alias{order_codons}
\title{order_codons}
\usage{
order_codons(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{string of nucleotides}

\item{start}{}
}
\value{
codons the RNA sequence subdivided into codons
}
\description{
order_codons
}
\examples{
input "ACTGCTA", output "ACT" "GCT"
}
