% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_to_u.R
\name{t_to_u}
\alias{t_to_u}
\title{t_to_u}
\usage{
t_to_u(dna_seq)
}
\arguments{
\item{dna_seq}{with nucleotides A, C, G and T}
}
\value{
rna_seq with nucleotides A, C, G and U
}
\description{
t_to_u
}
\examples{
input "ATTCG", get "AUUCG"
}
