% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_rna.R
\name{translate_rna}
\alias{translate_rna}
\title{translate_rna}
\usage{
translate_rna(codons)
}
\arguments{
\item{codons}{ordered RNA sequence divided into 3-nucleotide long subsequences}
}
\value{
aa_seq amino acid sequence translated from the RNA sequence
}
\description{
translate_rna
}
\examples{
AUG to met
}
